Estimation Engine: Refactor datasets
